requirements.txt
customtkinter>=5.1.2
Pillow>=9.0.0
flask>=2.0.0
requests>=2.25.0
pyinstaller>=5.0.0
websockets>=10.0.0